# VibeCoreWeb/ClientApp/src/api/models/problemDetails.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * VibeCore API
 * OpenAPI spec version: v1
 */

export interface ProblemDetails {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  status?: number | null;
  /** @nullable */
  detail?: string | null;
  /** @nullable */
  instance?: string | null;
  [key: string]: unknown;
}