# VibeCoreWeb/ClientApp/src/api/models/scheduledTaskDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * VibeCore API
 * OpenAPI spec version: v1
 */
import type { ScheduledTaskKind } from './scheduledTaskKind.ts';

export interface ScheduledTaskDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  handlerKey?: string | null;
  kind?: ScheduledTaskKind;
  /** @nullable */
  cronExpression?: string | null;
  /** @nullable */
  timeZoneId?: string | null;
  /** @nullable */
  runAt?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  nextFireAt?: string | null;
  /** @nullable */
  previousFireAt?: string | null;
}