# VibeCoreWeb/ClientApp/src/api/models/scheduledTaskHandlerDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * VibeCore API
 * OpenAPI spec version: v1
 */

export interface ScheduledTaskHandlerDto {
  /** @nullable */
  key?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  description?: string | null;
  retryCount?: number;
  allowsConcurrentExecution?: boolean;
}