# VibeCoreWeb/ClientApp/src/api/models/scheduledTaskKind.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * VibeCore API
 * OpenAPI spec version: v1
 */

export type ScheduledTaskKind = typeof ScheduledTaskKind[keyof typeof ScheduledTaskKind];


export const ScheduledTaskKind = {
  Cron: 'Cron',
  OneTime: 'OneTime',
} as const;