# VibeCoreWeb/ClientApp/src/api/models/scheduledTaskRunDto.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * VibeCore API
 * OpenAPI spec version: v1
 */

export interface ScheduledTaskRunDto {
  id?: string;
  scheduleId?: string;
  /** @nullable */
  handlerKey?: string | null;
  /** @nullable */
  status?: string | null;
  attempt?: number;
  startedAt?: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  durationMilliseconds?: number | null;
  /** @nullable */
  errorSummary?: string | null;
}